server/requirements.txt
aiohttp==3.8.5
aiosqlite==0.19.0
python-socketio==5.8.0
pytest==7.4.0
pytest-mock==3.11.1
pytest-aiohttp==1.0.4
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.0.0
mypy==1.4.1
isort==5.12.0